# Conflicts:
#	example/PluginStarterTestApplications/Intent_test/lib/main.dart
#	example/PluginStarterTestApplications/openthis/lib/main.dart
#	lib/src/communication/communication_api.dart

+++ lib/src/communication/storage_wrapper/owner_enforcer.dart
@@ -6,31 +6,38 @@ class OwnerEnforcerWrapper extends StorageController {
   final StorageController _controller;
   final String _owner;
   final bool _enforceTimestampUpdate;
+  final bool _pulledFromFactory;
 
   OwnerEnforcerWrapper(this._controller, this._owner,
-      [this._enforceTimestampUpdate = true]);
+      [this._enforceTimestampUpdate = true, this._pulledFromFactory = false]);
 
   String get owner => _owner;
 
   bool get enforceTimestampUpdate => _enforceTimestampUpdate;
 
+  bool get pulledFromFactory => _pulledFromFactory;
+
+  Node _updateOwnerAndTimestamp(Node n) {
+    if (n.owner != owner) {
+      (n as NodeImpl).set(Field.owner, owner);
+      if (_enforceTimestampUpdate) {
+        n.touch();
+      }
+    }
+    return n;
+  }
+
   @override
   Future<void> add(Node node) async {
     Node n = await node.deepClone();
-    n.owner = _owner;
-    if (_enforceTimestampUpdate) {
-      n.touch();
-    }
+    n = _updateOwnerAndTimestamp(n);
     return await _controller.add(n);
   }
 
   @override
   Future<bool> addOrUpdate(Node node) async {
     Node n = await node.deepClone();
-    n.owner = _owner;
-    if (_enforceTimestampUpdate) {
-      n.touch();
-    }
+    n = _updateOwnerAndTimestamp(n);
     return await _controller.addOrUpdate(n);
   }
 
@@ -39,7 +46,9 @@ class OwnerEnforcerWrapper extends StorageController {
     if (_enforceTimestampUpdate) {
       value.touch();
     }
-    return await _controller.addValue(path, value);
+    if ((await _controller.getNodeOrTombstone(path)).owner == _owner) {
+      return await _controller.addValue(path, value);
+    }
   }
 
   @override
@@ -52,7 +61,11 @@ class OwnerEnforcerWrapper extends StorageController {
   Future<Node> delete(String path) async {
     // nothing to do
     // timestamp is updated anyway
-    return await _controller.delete(path);
+    if ((await _controller.getNodeOrTombstone(path)).owner == _owner) {
+      return await _controller.delete(path);
+    }
+    throw StorageException(
+        'cannot delete nodes of foreign owners (path: $path)');
   }
 
   @override
@@ -77,50 +90,80 @@ class OwnerEnforcerWrapper extends StorageController {
   @override
   Future<Node> get(String path) async {
     // nothing to do
-    return await _controller.get(path);
+    Node node = await _controller.get(path);
+    if (node.visibility == Visibility.green ||
+        node.visibility == Visibility.white ||
+        node.owner == _owner) {
+      return node;
+    }
+    throw StorageException('access denied');
   }
 
   @override
   Future<Node> getNodeOrTombstone(String path) async {
-    // nothing to do
-    return await _controller.getNodeOrTombstone(path);
+    Node node = await _controller.getNodeOrTombstone(path);
+    if (node.visibility == Visibility.green ||
+        node.visibility == Visibility.white ||
+        node.owner == _owner) {
+      return node;
+    }
+    throw StorageException('access denied');
   }
 
   @override
   Future<NodeValue?> getValue(String path, String key) async {
     // nothing to do
-    return await _controller.getValue(path, key);
+    Node node = await _controller.get(path);
+    if (node.visibility == Visibility.green ||
+        node.visibility == Visibility.white ||
+        node.owner == _owner) {
+      return await _controller.getValue(path, key);
+    }
+    throw StorageException('access denied');
   }
 
   @override
   Future<void> registerChangeListener(
       StorageListener listener, SearchCriteria criteria) async {
     // nothing to do
+    criteria.owner = owner;
     return await _controller.registerChangeListener(listener, criteria);
   }
 
   @override
   Future<void> rename(String oldPath, String newName) async {
     // nothing to do
-    return await _controller.rename(oldPath, newName);
+    Node node = await _controller.get(oldPath);
+    if (node.owner == _owner) {
+      return await _controller.rename(oldPath, newName);
+    }
+    throw StorageException('access denied');
   }
 
   @override
   Future<List<Node>> search(SearchCriteria criteria) async {
     // nothing to do
+    criteria.owner = owner;
     return await _controller.search(criteria);
   }
 
   @override
   Future<void> update(Node node) async {
     // nothing to do
-    return await _controller.update(node);
+    node = _updateOwnerAndTimestamp(node);
+    Node nodeOld = await _controller.get(node.path);
+    if (nodeOld.owner == _owner) {
+      return await _controller.update(node);
+    }
+    throw StorageException('access denied');
   }
 
   @override
   Future<void> updateValue(String nodeName, NodeValue value) async {
-    // nothing to do
-    return await _controller.updateValue(nodeName, value);
+    if ((await _controller.get(nodeName)).owner == _owner) {
+      return await _controller.updateValue(nodeName, value);
+    }
+    throw StorageException('access denied');
   }
 
   @override
@@ -132,11 +175,15 @@ class OwnerEnforcerWrapper extends StorageController {
   @override
   Future<bool> addOrUpdateValue(String path, NodeValue value) async {
     // nothing to do (Is that so?)
-    return await _controller.addOrUpdateValue(path, value);
+    if ((await _controller.get(path)).owner == _owner) {
+      return await _controller.addOrUpdateValue(path, value);
+    }
+    throw StorageException('access denied');
   }
 
   @override
   Future<String> dump([String rootNode = ':', String prefix = '']) async {
+    //TODO(mgwerder): filter search results of foreign owners
     return await _controller.dump(rootNode, prefix);
   }
 }
